samples/ds-ml-ai--data-science-machine-learning-artificial-intelligence/11-etl-extract-transform-load/extractions/text-from-documents/pdf/pdfminer.six/requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
pdfminer.six==20250506
pillow==12.0.0
pycparser==2.23
